requirements.txt
rich==12.2.0
# rospy==1.15.14
PyYAML==5.3.1
rospkg==1.4.0
spacy==3.3.0
tabulate==0.8.9
scipy==1.8.0
plotly==5.8.0
pydash==5.1.0
pandas==1.4.2
openpyxl==3.0.10
numpy==1.22.3
matplotlib==3.5.2
google-cloud-dialogflow
graphviz
black